@@ -1,25 +1,35 @@
 {
-    "name": "light/hashids",
-    "description": "Hashids for Yii2",
-    "keywords": ["hashids", "yii2", "hash", "hashid", "ids", "youtube", "biyly", "encrypt", "decrypt"],
-    "require": {
-      "php": ">=5.4.0",
-      "yiisoft/yii2": "2.0.*",
-      "hashids/hashids": "1.0.5"
-    },
-    "license": "MIT",
-    "authors": [
-      {
-        "name": "lichunqiang",
-        "email": "light-li@hotmail.com"
-      }
-    ],
-    "autoload": {
-      "psr-4": {
-        "light\\hashids\\": "src/"
-      }
-    },
-    "require-dev": {
-        "phpunit/phpunit": "^4.7"
+  "name": "light/hashids",
+  "description": "Hashids for Yii2",
+  "keywords": ["hashids", "yii2", "hash", "hashid", "ids", "youtube", "biyly", "encrypt", "decrypt"],
+  "require": {
+    "php": "^5.6.4 || ^7.0",
+    "yiisoft/yii2": "~2.0.14",
+    "hashids/hashids": "^2.0"
+  },
+  "license": "MIT",
+  "authors": [
+    {
+      "name": "lichunqiang",
+      "email": "light-li@hotmail.com"
     }
+  ],
+  "autoload": {
+    "psr-4": {
+      "light\\hashids\\": "src/"
+    }
+  },
+  "autoload-dev": {
+    "psr-4": {
+      "light\\tests\\": "tests/"
+    }
+  },
+  "require-dev": {
+      "phpunit/phpunit": "^6.0"
+  },
+  "extra": {
+    "branch-alias": {
+      "dev-master": "2.0.x-dev"
+    }
+  }
 }
